Conflicts:
	compiler/cli/src/org/jetbrains/jet/cli/js/K2JSCompiler.java
	idea/src/org/jetbrains/jet/plugin/project/IDEAConfig.java
	idea/src/org/jetbrains/jet/plugin/project/K2JSModuleComponent.java
	idea/src/org/jetbrains/jet/plugin/quickfix/JsModuleSetUp.java
	js/js.tests/test/org/jetbrains/k2js/test/semantics/InvokeMethodTest.java
	js/js.translator/src/org/jetbrains/k2js/analyze/AnalyzerFacadeForJS.java
	js/js.translator/src/org/jetbrains/k2js/analyze/JsConfiguration.java
	js/js.translator/src/org/jetbrains/k2js/config/ClassPathLibraryDefintionsConfig.java
	js/js.translator/src/org/jetbrains/k2js/config/LibrarySourceDirectoriesConfig.java
	js/js.translator/src/org/jetbrains/k2js/config/LibrarySourcesConfig.java
	js/js.translator/src/org/jetbrains/k2js/generate/CodeGenerator.java
	js/js.translator/src/org/jetbrains/k2js/translate/context/StandardClasses.java
	js/js.translator/src/org/jetbrains/k2js/translate/context/TraceableThisAliasProvider.java
	js/js.translator/src/org/jetbrains/k2js/translate/context/generator/Generator.java
	js/js.translator/src/org/jetbrains/k2js/translate/intrinsic/functions/basic/CallStandardMethodIntrinsic.java
	js/js.translator/src/org/jetbrains/k2js/translate/intrinsic/functions/factories/PrimitiveBinaryOperationFIF.java
	js/js.translator/src/org/jetbrains/k2js/translate/intrinsic/functions/patterns/PatternBuilder.java
	js/js.translator/src/org/jetbrains/k2js/translate/reference/CallType.java
	js/js.translator/src/org/jetbrains/k2js/translate/utils/FunctionBodyTranslator.java
	js/js.translator/src/org/jetbrains/k2js/translate/utils/closure/CaptureClosureVisitor.java
	js/js.translator/src/org/jetbrains/k2js/translate/utils/closure/ClosureContext.java
	js/js.translator/src/org/jetbrains/k2js/translate/utils/closure/ClosureUtils.java

+++ compiler/frontend/src/org/jetbrains/jet/lang/resolve/constants/CompileTimeConstantResolver.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2012 JetBrains s.r.o.
+ * Copyright 2010-2013 JetBrains s.r.o.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -132,7 +132,7 @@ public class CompileTimeConstantResolver {
         }
         else if (JetTypeChecker.INSTANCE.isSubtypeOf(builtIns.getFloatType(), expectedType)) {
             try {
-                return new DoubleValue(Float.parseFloat(text));
+                return new FloatValue(Float.parseFloat(text));
             }
             catch (NumberFormatException e) {
                 return OUT_OF_RANGE;
